requirements.txt
vk_api==11.9.9
SQLAlchemy==2.0.9
psycopg2==2.9.6
python-dotenv==0.21.1
python_dateutil==2.8.2